proyecto Hibernate terminado. Pruebas OK

+++ Servidor/branches/PruebasHibernateCasino/src/Beans/Mesas.java
@@ -2,7 +2,6 @@ package Beans;
 // Generated 09-mar-2009 10:54:47 by Hibernate Tools 3.2.1.GA
 
 
-import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -12,19 +11,19 @@ import java.util.Set;
 public class Mesas  implements java.io.Serializable {
 
 
-     private BigDecimal codigo;
+     private int codigo;
      private Salas salas;
-     private BigDecimal puestos;
-     private BigDecimal jugadores;
-     private BigDecimal apuestamin;
-     private BigDecimal apuestamax;
+     private int puestos;
+     private int jugadores;
+     private double apuestamin;
+     private double apuestamax;
      private Set partidases = new HashSet(0);
 
     public Mesas() {
     }
 
 	
-    public Mesas(BigDecimal codigo, Salas salas, BigDecimal puestos, BigDecimal jugadores, BigDecimal apuestamin, BigDecimal apuestamax) {
+    public Mesas(int codigo, Salas salas, int puestos, int jugadores, double apuestamin, double apuestamax) {
         this.codigo = codigo;
         this.salas = salas;
         this.puestos = puestos;
@@ -32,7 +31,7 @@ public class Mesas  implements java.io.Serializable {
         this.apuestamin = apuestamin;
         this.apuestamax = apuestamax;
     }
-    public Mesas(BigDecimal codigo, Salas salas, BigDecimal puestos, BigDecimal jugadores, BigDecimal apuestamin, BigDecimal apuestamax, Set partidases) {
+    public Mesas(int codigo, Salas salas, int puestos, int jugadores, double apuestamin, double apuestamax, Set partidases) {
        this.codigo = codigo;
        this.salas = salas;
        this.puestos = puestos;
@@ -42,11 +41,11 @@ public class Mesas  implements java.io.Serializable {
        this.partidases = partidases;
     }
    
-    public BigDecimal getCodigo() {
+    public int getCodigo() {
         return this.codigo;
     }
     
-    public void setCodigo(BigDecimal codigo) {
+    public void setCodigo(int codigo) {
         this.codigo = codigo;
     }
     public Salas getSalas() {
@@ -56,32 +55,32 @@ public class Mesas  implements java.io.Serializable {
     public void setSalas(Salas salas) {
         this.salas = salas;
     }
-    public BigDecimal getPuestos() {
+    public int getPuestos() {
         return this.puestos;
     }
     
-    public void setPuestos(BigDecimal puestos) {
+    public void setPuestos(int puestos) {
         this.puestos = puestos;
     }
-    public BigDecimal getJugadores() {
+    public int getJugadores() {
         return this.jugadores;
     }
     
-    public void setJugadores(BigDecimal jugadores) {
+    public void setJugadores(int jugadores) {
         this.jugadores = jugadores;
     }
-    public BigDecimal getApuestamin() {
+    public double getApuestamin() {
         return this.apuestamin;
     }
     
-    public void setApuestamin(BigDecimal apuestamin) {
+    public void setApuestamin(double apuestamin) {
         this.apuestamin = apuestamin;
     }
-    public BigDecimal getApuestamax() {
+    public double getApuestamax() {
         return this.apuestamax;
     }
     
-    public void setApuestamax(BigDecimal apuestamax) {
+    public void setApuestamax(double apuestamax) {
         this.apuestamax = apuestamax;
     }
     public Set getPartidases() {
